+++ frontend/src/pages/ToysByCompany.jsx
@@ -3,7 +3,7 @@ import axios from 'axios';
 import { endpoints } from '../endpoints/Endpoints';
 
 import ToysByCompanyContent from '../components/content/ToysByCompanyContent';
-import { Pagination } from 'react-bootstrap';
+import { Form, Pagination } from 'react-bootstrap';
 
 const ToysByCompany = () => {
   const [toys, setToys] = useState([]);
@@ -14,22 +14,58 @@ const ToysByCompany = () => {
   const [allTotalQuantity, setAllTotalQuantity] = useState(0);
   const [allTotalPrice, setAllTotalPrice] = useState(0);
 
+  const [filterOptions, setFilterOptions] = useState({
+    companies: [],
+    brands: [],
+    series: [],
+    collections: []
+  });
+
+  const [selectedFilters, setSelectedFilters] = useState({
+    company: '',
+    brand: '',
+    series: '',
+    collection: ''
+  });
+
+  const [filteredToys, setFilteredToys] = useState([]);
+
   useEffect(() => {
     axios.get(endpoints.API_URL + 'toys')
       .then((response) => {
         const sortedToys = response.data.sort((a, b) => {
           const companyComparison = a.company.localeCompare(b.company);
           if (companyComparison !== 0) return companyComparison;
-  
+
           const brandComparison = a.brand.localeCompare(b.brand);
           if (brandComparison !== 0) return brandComparison;
-  
+
           const seriesComparison = a.series.localeCompare(b.series);
           if (seriesComparison !== 0) return seriesComparison;
-  
+
           return a.collection.localeCompare(b.collection);
         });
-      setToys(sortedToys);
+        setToys(sortedToys);
+
+        // Extract filter options
+        const companies = [...new Set(sortedToys.map(toy => toy.company))];
+        const brands = [...new Set(sortedToys.map(toy => toy.brand))];
+        const series = [...new Set(sortedToys.map(toy => toy.series))];
+        const collections = [...new Set(sortedToys.map(toy => toy.collection))];
+
+        // Sort filter options alphabetically
+        companies.sort();
+        brands.sort();
+        series.sort();
+        collections.sort();
+
+        // Update filter options in state
+        setFilterOptions({
+          companies,
+          brands,
+          series,
+          collections
+        });
 
         // Calculate total price and quantity for all toys
         let totalQuantity = 0;
@@ -47,9 +83,21 @@ const ToysByCompany = () => {
       });
   }, []);
 
+  useEffect(() => {
+    const filtered = toys.filter(toy => (
+      (!selectedFilters.company || toy.company === selectedFilters.company) &&
+      (!selectedFilters.brand || toy.brand === selectedFilters.brand) &&
+      (!selectedFilters.series || toy.series === selectedFilters.series) &&
+      (!selectedFilters.collection || toy.collection === selectedFilters.collection) &&
+      (!selectedFilters.completed || toy.completed === selectedFilters.completed)
+    ));
+
+    setFilteredToys(filtered);
+  }, [toys, selectedFilters]);
+
   // Calculate the range of pages to display
   const pageRange = 8;
-  const totalPages = Math.ceil(toys.length / toysPerPage);
+  const totalPages = Math.ceil(filteredToys.length / toysPerPage);
 
   const startPage = Math.max(currentPage - Math.floor(pageRange / 2), 1);
   const endPage = Math.min(startPage + pageRange - 1, totalPages);
@@ -57,10 +105,14 @@ const ToysByCompany = () => {
   // Get current toys for the current page
   const indexOfLastToy = currentPage * toysPerPage;
   const indexOfFirstToy = indexOfLastToy - toysPerPage;
-  const currentToys = toys.slice(indexOfFirstToy, indexOfLastToy);
+  const currentToys = filteredToys.slice(indexOfFirstToy, indexOfLastToy);
 
   // Change page
   const paginate = (pageNumber) => setCurrentPage(pageNumber);
+
+  useEffect(() => {
+    setCurrentPage(1);
+  }, [selectedFilters]);
 
   // Calculate the total price for the currently displayed toys
   let totalQuantity = 0;
@@ -72,12 +124,149 @@ const ToysByCompany = () => {
 
   return (
     <>
+      <div className="filter-section">
+        <div className="row">
+          <div className="col">
+            <Form.Select
+              size="sm"
+              aria-label="Companies"
+              value={selectedFilters.company}
+              onChange={(e) => {
+                const selectedCompany = e.target.value;
+                const selectedBrand = selectedFilters.brand;
+
+                if (selectedCompany === '') {
+                  // Reset all filters when "All Companies" is selected
+                  setSelectedFilters({
+                    company: '',
+                    brand: '',
+                    series: '',
+                    collection: ''
+                  });
+                } else {
+                  setSelectedFilters({
+                    ...selectedFilters,
+                    company: selectedCompany,
+                    brand: selectedCompany !== selectedFilters.company ? '' : selectedBrand,
+                  });
+                }
+              }}
+            >
+              <option value="">All Companies</option>
+              {filterOptions.companies.map((company) => (
+                <option key={company} value={company}>
+                  {company}
+                </option>
+              ))}
+            </Form.Select>
+          </div>
+          <div className="col">
+            <Form.Select
+              size="sm"
+              aria-label="Brands" value={selectedFilters.brand}
+              onChange={(e) => setSelectedFilters({ ...selectedFilters, brand: e.target.value })}
+            >
+              <option value="">All Brands</option>
+              {filterOptions.brands
+                .filter((brand) => !selectedFilters.company || toys.some((toy) => toy.company === selectedFilters.company && toy.brand === brand))
+                .map((brand) => (
+                  <option key={brand} value={brand}>
+                    {brand}
+                  </option>
+                ))}
+            </Form.Select>
+          </div>
+          <div className="col">
+            <Form.Select
+              size="sm"
+              aria-label="Series"
+              value={selectedFilters.series}
+              onChange={(e) => {
+                const selectedSeries = e.target.value;
+                setSelectedFilters({ ...selectedFilters, series: selectedSeries });
+              }}
+            >
+              <option value="">All Series</option>
+              {filterOptions.series
+                .filter((series) =>
+                  !selectedFilters.company ||
+                  !selectedFilters.brand ||
+                  toys.some(
+                    (toy) =>
+                      toy.company === selectedFilters.company &&
+                      toy.brand === selectedFilters.brand &&
+                      toy.series === series
+                  )
+                )
+                .map((series) => (
+                  // Filter out blank/empty series
+                  series && (
+                    <option key={series} value={series}>
+                      {series}
+                    </option>
+                  )
+                ))}
+            </Form.Select>
+          </div>
+          <div className="col">
+            <Form.Select
+              size="sm"
+              aria-label="Collections"
+              value={selectedFilters.collection}
+              onChange={(e) => {
+                const selectedCollection = e.target.value;
+                setSelectedFilters({ ...selectedFilters, collection: selectedCollection });
+              }}
+            >
+              <option value="">All Collections</option>
+              {filterOptions.collections
+                .filter((collection) =>
+                  !selectedFilters.company ||
+                  !selectedFilters.brand ||
+                  toys.some(
+                    (toy) =>
+                      toy.company === selectedFilters.company &&
+                      toy.brand === selectedFilters.brand &&
+                      toy.series === selectedFilters.series &&
+                      toy.collection === collection
+                  )
+                )
+                .map((collection) => (
+                  // Filter out blank/empty collections
+                  collection && (
+                    <option key={collection} value={collection}>
+                      {collection}
+                    </option>
+                  )
+                ))}
+            </Form.Select>
+          </div>
+          <div className="col">
+            <button
+              className="btn btn-primary btn-sm"
+              onClick={() => {
+                // Clear all selected filters
+                setSelectedFilters({
+                  company: '',
+                  brand: '',
+                  series: '',
+                  collection: '',
+                  completed: '',
+                });
+              }}
+            >
+              Clear Filters
+            </button>
+          </div>
+        </div>
+      </div>
+
       <div className="page-title">
         Toys by Company
 
         <div className="totals">
-          <div>Collection Price: ${allTotalPrice.toFixed(2)}</div>
-          <div className="current-page">Current Page Price: ${totalPrice.toFixed(2)}</div>
+          <div>Collection Total Value: ${allTotalPrice.toFixed(2)}</div>
+          <div className="current-page">Current Page Value: ${totalPrice.toFixed(2)}</div>
         </div>
       </div>
 
@@ -86,7 +275,6 @@ const ToysByCompany = () => {
         toysPerPage={toysPerPage}
         currentToys={currentToys}  // Pass the current toys to the child component
         toys={toys} // Pass all toys to the child component
-        //paginate={paginate} // Pass the paginate function to the child component
       />
 
       <div className="pagination-wrapper">
